glr.cc: shy away from unsigned types
On my machine, `yystates.begin() + newsize` wants a `long int` on the
RHS, and warns about size_t.  Let's rather use signed types for sizes.
This is a general trend in C/C++: avoid unsigned types (unless for
bitfields, etc.).  And it matches what was done in yacc.c and lalr1.cc.